Introduce GenImplicit
