+++ src/CMakeLists.txt
@@ -86,6 +86,7 @@ set(SOURCES
         irgenerator/GenBuiltinFunctions.cpp
         irgenerator/GenExpressions.cpp
         irgenerator/GenValues.cpp
+        irgenerator/GenImplicit.cpp
         irgenerator/StdFunctionManager.cpp
         irgenerator/StdFunctionManager.h
         irgenerator/OpRuleConversionManager.cpp
